implement precompiled headers for baron and jni so that we can split android.cpp into multiple files
