+++ thunks/libc/impl_header.h
@@ -1,6 +1,8 @@
 // THIS FILE IS AUTOMATICALLY GENERATED
 // PLEASE CHECK OUT `generate_libc.py` TO REGENERATE IT
+#ifndef _GNU_SOURCE
 #define _GNU_SOURCE
+#endif
 #include <dlfcn.h>
